+++ src/dynamic_columns/exceptions.py
@@ -1,0 +1,2 @@
+class CodeAccessNotAllowed(Exception):
+    pass