+++ HISTORY.rst
@@ -19,6 +19,7 @@ dev
 * poprawki aplikacji do uruchamiania procesów w tle (b/n),
 * nie wyświetlaj przycisku "pokaż w PBN" gdy autor nie ma określonego odpowiednika w PBN (b/n),
 * szybsze wyświetlanie listy nazwisk dla odpowiedników PBN dla autora (b/n),
+* możliwość wyboru widocznych kolumn w module redagowania (b/n),
 
 202209.1088
 -----------